Moved HAVING clause to SelectQuery
