+++ src/Query.php
@@ -8,7 +8,6 @@ use Girgias\QueryBuilder\Exceptions\InvalidSqlColumnNameException;
 use Girgias\QueryBuilder\Exceptions\InvalidSqlFieldNameException;
 use Girgias\QueryBuilder\Exceptions\InvalidSqlTableNameException;
 use Girgias\QueryBuilder\Exceptions\UnexpectedSqlOperatorException;
-use Girgias\QueryBuilder\Exceptions\UnexpectedSqlFunctionException;
 use InvalidArgumentException;
 use RuntimeException;
 use TypeError;
@@ -34,11 +33,6 @@ abstract class Query
      * @var ?array<int, string>
      */
     protected $where;
-
-    /**
-     * @var ?array<int, string>
-     */
-    protected $having;
 
     /**
      * @var ?array<string, string>
@@ -181,73 +175,6 @@ abstract class Query
         }
 
         $this->where[] = $column . ' NOT BETWEEN ' . $this->buildBetweenSqlString($start, $end);
-
-        return $this;
-    }
-
-    /**
-     * Add a HAVING clause to the Query
-     *
-     * @param string $column
-     * @param string $aggregateFunction
-     * @param string $operator
-     * @param int $conditionValue
-     * @return Query
-     */
-    final public function having(string $column, string $aggregateFunction, string $operator, int $conditionValue): self
-    {
-        if (!$this->isValidSqlName($column)) {
-            throw new InvalidSqlColumnNameException('HAVING', $column);
-        }
-
-        if (!AggregateFunctions::isValidValue($aggregateFunction)) {
-            throw new UnexpectedSqlFunctionException('HAVING', $aggregateFunction);
-        }
-
-        if (!SqlOperators::isValidValue($operator)) {
-            throw new UnexpectedSqlOperatorException('HAVING', $operator);
-        }
-
-        $this->having[] = $aggregateFunction . '(' . $column . ') ' . $operator . ' ' . $conditionValue;
-
-        return $this;
-    }
-
-    /**
-     * Add a HAVING clause to the Query which should be ORed with the previous use of a HAVING clause
-     *
-     * @param string $column
-     * @param string $aggregateFunction
-     * @param string $operator
-     * @param int $conditionValue
-     * @return Query
-     */
-    final public function havingOr(
-        string $column,
-        string $aggregateFunction,
-        string $operator,
-        int $conditionValue
-    ): self {
-        if (!$this->isValidSqlName($column)) {
-            throw new InvalidSqlColumnNameException('HAVING', $column);
-        }
-
-        if (!AggregateFunctions::isValidValue($aggregateFunction)) {
-            throw new UnexpectedSqlFunctionException('HAVING', $aggregateFunction);
-        }
-
-        if (!SqlOperators::isValidValue($operator)) {
-            throw new UnexpectedSqlOperatorException('HAVING', $operator);
-        }
-
-        if (is_null($this->having)) {
-            throw new RuntimeException(
-                'Need to define at least another HAVING clause before utilizing havingOr method'
-            );
-        }
-
-        $this->having[] = '(' . array_pop($this->having) . ' OR ' .
-            $aggregateFunction . '(' . $column . ') ' . $operator . ' ' . $conditionValue . ')';
 
         return $this;
     }
